tagForGlobalCoarsening added.
Also change in splitGhost method.
This method now becomes a GhostChildrenInfo which is filled with the
children info.

Furthermore, a dummy class for hasFace3 and 4 was implemented, such that
by default the hface nb class is saet to this value. This default class
just does nothing. But with that ghost cells can be used without fake
boundary methods.


git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/alugrid/trunk@681 0d966ed9-3843-0410-af09-ebfb50bd7c74