Make configure abort if external ALUGrid was found.
This seems to be the only liable solution as using the old
ALUGrid via the dune-alugrid bindings will break eventually.
Most people will probably not notice a warning anyway.
In addition we give a hint to use --without-alugrid as there
might be installed ALUGrid versions in the default installation
path.

Signed-off-by: dedner <a.s.dedner@warwick.ac.uk>