use decltype magic to protect loadBalance( LBWeights ) from being used for a CommDataHandleIF
Because the CommDataHandleIF exports the DataType, it is indeed possible
to check whether a type T is derived from
(*)    CommDataHandle< T, typenanme T::DataType >.
We use this information to disable the generic method loadBalance( LBWeights )
in this case.

Note: If someone passes a type U derived from a type T derived from (*),
there is no way of checking this. In this case, the wrong method will be
used, hopefully resulting in a compile time error.