periodic elements seem to work now in 2d
the problem was that the search for the periodic neighbour deletes the periodic
boundary faces from the boundary and the elements get oriented between insertion
and this. So it does not find the elements. Now the search is done after the recreation of
the boundaryIds with respect to the new element orientation


Former-commit-id: d07de1bc2b4684af34ef591a0d226d7eb245db58