[bugfix] don't fail without SIONlib
The backup-restore example failed if SIONlib is not found, because the
corresponding read/write methods silently do nothing. This causes the grid to
be restored from an empty stream, resulting in an error.

This patch fixes the issue by restoring the grid from the SIONlib file only if
SIONlib has been found. Otherwise, this option of restoring is silently ignored.


Former-commit-id: 1f946dae1358249b9c231c605ec1d96013ed60af