separate topology traits data and accessor functions
Original intent was to get rid of clang warnings like this:

  dune/dune-alugrid/dune/alugrid/3d/topology.hh:240:12: warning:
  instantiation of variable
  'Dune::ElementTopologyMapping<Dune::tetra>::generic2aluFace_' required
  here, but no definition is available [-Wundefined-var-template]

These occured because the definitions were provided in a separate
translation unit `3d/topology.cc` and not known at time of
instantiation.
The patch moves traits data to a seperate templated traits class and
lets the mapping class inherit from it. Better readability is a bonus.


Former-commit-id: 68ec62acae3b180b6164964aa21ca995cafbe5d9