Add a workaround to fix compilation with clang 6
clang 6 hangs indefinitely when trying to compile gitter_mgb.cc or gitter_pll_mgb.cc with an
optimization level of -O2 or higher. This patch adds some CMake logic that detects the problematic
case and forces the optimization level for those two files down to -O1 if necessary, but avoids
bumping it up in case of a debug build.