[CMake] Find master dune-grid in home dir even with installed stable.
When not using dunecontrol but pure cmake, an installed dune-grid was
preferred. In my case I was testing master and had 2.6 installed. The
latter was found when running cmake directly.

This change fixes this by introducing the same trickery as for
dune-common. Please note that this is not needed for modules that
adhere to the usual workflow (without an explicit search for
dependencies and relying on DuneMacros). But here a different approach
was used and that probably has its reasons.